float edit

+++ python_src/server_for_motor.py
@@ -25,8 +25,8 @@ class RobotControlServer:
         speed_right = int(command[1]) 
         robot.set_speed_cms_left(speed_left)
         robot.set_speed_cms_right(speed_right)
-        left_const = int(command[2])
-        right_const = int(command[3])
+        left_const = float(command[2])
+        right_const = float(command[3])
         robot.set_consts(left_const,right_const)
 
     def start(self):
